+++ app/(payload)/admin/importMap.js
@@ -1,6 +1,8 @@
+import { default as default_c4b34a96eb3c9f20c6bc285cd98d312b } from '@/components/auth/AdminGoogleLoginBefore'
 import { CollectionCards as CollectionCards_f9c02e79a4aed9a3924487c0cd4cafb1 } from '@payloadcms/next/rsc'
 
 /** @type import('payload').ImportMap */
 export const importMap = {
+  "@/components/auth/AdminGoogleLoginBefore#default": default_c4b34a96eb3c9f20c6bc285cd98d312b,
   "@payloadcms/next/rsc#CollectionCards": CollectionCards_f9c02e79a4aed9a3924487c0cd4cafb1
 }